refactor: 서비스간 순환참조가 발생하지 않도록 별도의 계층 생성
