+++ src/main/java/com/flab/readnshare/domain/review/service/ReviewService.java
@@ -1,11 +1,7 @@
 package com.flab.readnshare.domain.review.service;
 
-import com.flab.readnshare.domain.book.domain.Book;
-import com.flab.readnshare.domain.book.dto.BookDto;
-import com.flab.readnshare.domain.book.service.BookService;
 import com.flab.readnshare.domain.member.domain.Member;
 import com.flab.readnshare.domain.review.domain.Review;
-import com.flab.readnshare.domain.review.dto.SaveReviewRequestDto;
 import com.flab.readnshare.domain.review.dto.UpdateReviewRequestDto;
 import com.flab.readnshare.domain.review.repository.ReviewRepository;
 import com.flab.readnshare.global.common.exception.ReviewException;
@@ -15,27 +11,21 @@ import org.springframework.transaction.annotation.Transactional;
 
 
 @Service
-@Transactional
+@Transactional(readOnly = true)
 @RequiredArgsConstructor
 public class ReviewService {
     private final ReviewRepository reviewRepository;
-    private final BookService bookService;
 
-    @Transactional(readOnly = true)
     public Review findById(Long reviewId) {
         return reviewRepository.findById(reviewId)
                 .orElseThrow(ReviewException.ReviewNotFoundException::new);
     }
 
-    public Long save(SaveReviewRequestDto dto, Member signInMember) {
-        BookDto bookDto = dto.getBook();
-        Book book = bookService.save(bookDto);
-
-        Review review = dto.toEntity(signInMember, book);
-
+    public Long save(Review review) {
         return reviewRepository.save(review).getId();
     }
 
+    @Transactional
     public Long update(Long reviewId, Member signInMember, UpdateReviewRequestDto dto) {
         Review review = findById(reviewId);
         review.verifyMember(signInMember);
@@ -45,6 +35,7 @@ public class ReviewService {
         return review.getId();
     }
 
+    @Transactional
     public void delete(Long reviewId, Member signInMember) {
         Review review = findById(reviewId);
         review.verifyMember(signInMember);
